# index-server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Dev-connector API",
    "description": "API description format for different endpoints"
  },
  "schemes": ["http", "https"],
  "host": "localhost:80",
  "paths": {
...
    "/api/users/register": {
      "post":
      {
        "summary": "Register User",
        "tags": ["Users"],
        "description": "Route for registering an user to the database",
        "parameters": [
          {
            "name": "RegUser",
            "in": "body",
            "description": "Register credentials",
            "schema": {
              "type": "object",
              "required": ["name", "email", "password", "password2"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "password2": {
                  "type": "string"
                }
              }
            }
          }],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "password2": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "The user has been successfully registered",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "format": "object"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "avatar": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "__v": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "summary": "Login User",
        "tags": ["Users"],
        "description": "Route to logging an user",
        "parameters": [{
          "name": "LogUser",
          "in": "body",
          "description": "Login credentials",
          "schema": {
            "required": ["email", "password"],
            "properties": {
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "The user has successfully logged in",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "The user can't be found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "JWT": {
      "description": "JSON WEB TOKEN",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [{
    "JWT": []
  }],
...
}